+++ common/setups/rasr/util/__init__.py
@@ -1,0 +1,3 @@
+from .rasr import *
+from .gmm import *
+from .nn import *
